requirements.txt
asgiref==3.8.1
Django==4.2.21
django-cors-headers==4.7.0
djangorestframework==3.16.0
sqlparse==0.5.3
typing_extensions==4.13.2